processing: preserve scroll position while zooming
This commit makes the Processing program preserve the scroll
position relative to timestamp, when the user is zooming in/out
the graph. Effectively, the left edge of the visible window
always represents the same timestamp, so that the user doesn't
have to search for the interesting region again after zooming.